+++ angular/projects/catalyst/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [14.4.0](https://github.com/haiilo/catalyst/compare/catalyst-angular-v14.3.0...catalyst-angular-v14.4.0) (2026-02-25)
+
+
+### Miscellaneous Chores
+
+* **catalyst-angular:** Synchronize catalyst versions
+
 ## [14.3.0](https://github.com/haiilo/catalyst/compare/catalyst-angular-v14.2.0...catalyst-angular-v14.3.0) (2026-02-10)